Add timeouts and workarounds
When executing the tests with Firefox, many tests fail due to the
pages taking a very long time to be fully loaded. The same failure
also happened with Chromium but more rarely.
Timeouts are not the best way to do wait for a page and hence
another solution should be found.